refactor: XMLDeserializationError to own file
