+++ Source/XMLDeserializationError.swift
@@ -1,0 +1,59 @@
+//
+//  XMLDeserializationError.swift
+//  SWXMLHash
+//
+//  Created by David Mohundro on 5/3/22.
+//
+
+import Foundation
+
+public enum XMLDeserializationError: Error, CustomStringConvertible {
+    case implementationIsMissing(method: String)
+    case nodeIsInvalid(node: String)
+    case nodeHasNoValue
+    case typeConversionFailed(type: String, element: XMLElement)
+    case attributeDoesNotExist(element: XMLElement, attribute: String)
+    case attributeDeserializationFailed(type: String, attribute: XMLAttribute)
+
+// swiftlint:disable identifier_name
+    @available(*, unavailable, renamed: "implementationIsMissing(method:)")
+    public static func ImplementationIsMissing(method: String) -> XMLDeserializationError {
+        fatalError("unavailable")
+    }
+    @available(*, unavailable, renamed: "nodeHasNoValue(_:)")
+    public static func NodeHasNoValue(_: IndexOps) -> XMLDeserializationError {
+        fatalError("unavailable")
+    }
+    @available(*, unavailable, renamed: "typeConversionFailed(_:)")
+    public static func TypeConversionFailed(_: IndexingError) -> XMLDeserializationError {
+        fatalError("unavailable")
+    }
+    @available(*, unavailable, renamed: "attributeDoesNotExist(_:_:)")
+    public static func AttributeDoesNotExist(_ attr: String, _ value: String) throws -> XMLDeserializationError {
+        fatalError("unavailable")
+    }
+    @available(*, unavailable, renamed: "attributeDeserializationFailed(_:_:)")
+    public static func AttributeDeserializationFailed(_ attr: String, _ value: String) throws -> XMLDeserializationError {
+        fatalError("unavailable")
+    }
+// swiftlint:enable identifier_name
+
+    /// The text description for the error thrown
+    public var description: String {
+        switch self {
+        case .implementationIsMissing(let method):
+            return "This deserialization method is not implemented: \(method)"
+        case .nodeIsInvalid(let node):
+            return "This node is invalid: \(node)"
+        case .nodeHasNoValue:
+            return "This node is empty"
+        case let .typeConversionFailed(type, node):
+            return "Can't convert node \(node) to value of type \(type)"
+        case let .attributeDoesNotExist(element, attribute):
+            return "Element \(element) does not contain attribute: \(attribute)"
+        case let .attributeDeserializationFailed(type, attribute):
+            return "Can't convert attribute \(attribute) to value of type \(type)"
+        }
+    }
+}
+
